Implemented plotting superclass, inherited by sygma and omega. Transferred all plot-related functionality to this class.
