+++ omega.py
@@ -70,9 +70,10 @@ import random
 # Import the class inherited by SYGMA
 import sygma
 from chem_evol import *
+from plotting import *
 
 
-class omega( chem_evol ):
+class omega( chem_evol, plotting ):
 
     '''
     Input parameters (OMEGA)
@@ -1791,894 +1792,6 @@ class omega( chem_evol ):
 ######################## Here start the analysis methods ######################################
 ###############################################################################################
 
-
-
-    def plot_mass(self,fig=0,specie='C',source='all',norm=False,label='',shape='',marker='',color='',markevery=20,multiplot=False,return_x_y=False,fsize=[10,4.5],fontsize=14,rspace=0.6,bspace=0.15,labelsize=15,legend_fontsize=14):
-    
-        '''
-	 mass evolution (in Msun) of an element or isotope vs time.
-	
-
-        Parameters
-        ----------
-
-
-	specie : string
-             isotope or element name, in the form 'C' or 'C-12'
-	source : string
-             Specifies if yields come from
-	     all sources ('all'), including
-	     AGB+SN1a, massive stars. Or from
-	     distinctive sources:
-	     only agb stars ('agb'), only
-	     SN1a ('SN1a'), or only massive stars
-             ('massive')
-        norm : boolean
-             If True, normalize to current total ISM mass
-	label : string
-	     figure label
-	marker : string
-	     figure marker
-	shape : string
-	     line style	
-	color : string
-	     color of line
-        fig : string,float
-	     to name the plot figure       
- 	       
-        Examples
-	----------
-
-	>>> s.plot('C-12')
-
-        '''
-	yaxis=specie
-	if len(label)<1:
-		label=yaxis
-		if source=='agb':
-			label=yaxis+', AGB'
-		if source=='massive':
-			label=yaxis+', Massive'
-		if source=='sn1a':
-			label=yaxis+', SNIa'
-
-        #Reserved for plotting
-        if not return_x_y:
-            shape,marker,color=self.__msc(source,shape,marker,color)
-
-        x=self.history.age
-        self.x=x
-        y=[]
-        yields_evol_all=self.history.ism_elem_yield
-        if yaxis in self.history.elements:
-            if source == 'all':
-                yields_evol=self.history.ism_elem_yield
-            elif source =='agb':
-                yields_evol=self.history.ism_elem_yield_agb
-            elif source == 'sn1a':
-                yields_evol=self.history.ism_elem_yield_1a
-            elif source == 'massive':
-                yields_evol=self.history.ism_elem_yield_massive
-            idx=self.history.elements.index(yaxis)
-        elif yaxis in self.history.isotopes:
-            if source == 'all':
-                yields_evol=self.history.ism_iso_yield
-            elif source =='agb':
-                yields_evol=self.history.ism_iso_yield_agb
-            elif source == 'sn1a':
-                yields_evol=self.history.ism_iso_yield_1a
-            elif source == 'massive':
-                yields_evol=self.history.ism_iso_yield_massive
-            idx=self.history.isotopes.index(yaxis)
-        else:
-            print 'Isotope or element not available'
-            return 0
-        for k in range(0,len(yields_evol)):
-            if norm == False:
-                y.append(yields_evol[k][idx])
-            else:
-                y.append( yields_evol[k][idx]/yields_evol_all[k][idx])
-
-
-        x=x[1:]
-        y=y[1:]
-        if multiplot==True:
-                return x,y  
-
-        #Reserved for plotting
-        if not return_x_y:
-           plt.figure(fig, figsize=(fsize[0],fsize[1]))
-           plt.xscale('log')
-           plt.xlabel('log-scaled age [yrs]')
-           if norm == False:
-               plt.ylabel('yields [Msun]')
-               plt.yscale('log')
-           else:
-               plt.ylabel('(IMF-weighted) fraction of ejecta')
-        self.y=y
-
-        #If x and y need to be returned ...
-        if return_x_y:
-            return x, y
-
-        else:
-            plt.plot(x,y,label=label,linestyle=shape,marker=marker,color=color,markevery=markevery)
-            plt.legend()
-            ax=plt.gca()
-            self.__fig_standard(ax=ax,fontsize=fontsize,labelsize=labelsize,rspace=rspace, bspace=bspace,legend_fontsize=legend_fontsize)
-	    plt.xlim(self.history.dt,self.history.tend)	
-	    #return x,y
-	    self.save_data(header=['Age[yrs]',specie],data=[x,y])
-
-
-    def plot_massfrac(self,fig=2,xaxis='age',yaxis='O-16',source='all',norm='no',label='',shape='',marker='',color='',markevery=20,fsize=[10,4.5],fontsize=14,rspace=0.6,bspace=0.15,labelsize=15,legend_fontsize=14):
-
-        '''
-	Plots mass fraction of isotope or element
-	vs either time or other isotope or element.
-
-        Parameters
-        ----------
-	xaxis : string 
-            either 'age' for time
-	    or isotope name, in the form e.g. 'C-12'
-	yaxis : string
-            isotope name, in the same form as for xaxis
-
-	source : string
-            Specifies if yields come from
-	    all sources ('all'), including
-	    AGB+SN1a, massive stars. Or from
-	    distinctive sources:
-	    only agb stars ('agb'), only
-	    SN1a ('SN1a'), or only massive stars
-            ('massive')
-
-        norm : string
-            if 'no', no normalization of mass fraction
-            if 'ini', normalization ot initial abundance
-        label : string
-             figure label
-        marker : string
-             figure marker
-        shape : string
-             line style 
-        color : string
-             color of line
-        fig : string,float
-             to name the plot figure       
-
-        
-        Examples
-        ----------
-
-	>>> s.plot_massfrac('age','C-12')
-
-        '''
-
-        import matplotlib.pyplot as plt
-
-        if len(label)<1:
-                label=yaxis
-
-	shape,marker,color=self.__msc(source,shape,marker,color)
-
-
-
-        plt.figure(fig, figsize=(fsize[0],fsize[1]))
-
- 
-        #Input X-axis
-        if '-' in xaxis:
-        #to test the different contributions
-	    if source == 'all':
-	       yields_evol=self.history.ism_iso_yield
-	    elif source =='agb':
-	      yields_evol=self.history.ism_iso_yield_agb
-	    elif source == 'sn1a':
-	      yields_evol=self.history.ism_iso_yield_1a
-	    elif source == 'massive':
-	     yields_evol=self.history.ism_iso_yield_massive
-	     iso_idx=self.history.isotopes.index(xaxis)
-            x=[]
-	    for k in range(1,len(yields_evol)):
-	       if norm=='no':
-	           x.append(yields_evol[k][iso_idx]/sum(yields_evol[k]))
-	       if norm=='ini':
-	           x.append(yields_evol[k][iso_idx]/sum(yields_evol[k])/yields_evol[0][iso_idx])
-            plt.xlabel('log-scaled X('+xaxis+')')
-            plt.xscale('log')
-        elif 'age' == xaxis:
-            x=self.history.age#[1:]
-            plt.xscale('log')
-            plt.xlabel('log-scaled Age [yrs]')
-        elif 'Z' == xaxis:
-            x=self.history.metallicity#[1:]
-            plt.xlabel('ISM metallicity')
-            plt.xscale('log')
-        elif xaxis in self.history.elements:
-            if source == 'all':
-                yields_evol=self.history.ism_elem_yield
-            elif source =='agb':
-                yields_evol=self.history.ism_elem_yield_agb
-            elif source == 'sn1a':
-                yields_evol=self.history.ism_elem_yield_1a
-            elif source == 'massive':
-                yields_evol=self.history.ism_elem_yield_massive
-            iso_idx=self.history.elements.index(xaxis)
-            x=[]
-            for k in range(1,len(yields_evol)):
-                if norm=='no':
-                    x.append(yields_evol[k][iso_idx]/sum(yields_evol[k]))
-                if norm=='ini':
-                    x.append(yields_evol[k][iso_idx]/sum(yields_evol[k])/yields_evol[0][iso_idx])
-                    print yields_evol[0][iso_idx]
-
-            plt.xlabel('log-scaled X('+xaxis+')')
-            plt.xscale('log')
-
-
-        #Input Y-axis
-        if '-' in yaxis:
-            #to test the different contributions
-            if source == 'all':
-                yields_evol=self.history.ism_iso_yield
-            elif source =='agb':
-                yields_evol=self.history.ism_iso_yield_agb
-            elif source == 'sn1a':
-                yields_evol=self.history.ism_iso_yield_1a
-            elif source == 'massive':
-                yields_evol=self.history.ism_iso_yield_massive
-            iso_idx=self.history.isotopes.index(yaxis)
-            y=[]
-            #change of xaxis array 'age 'due to continue statement below
-            if xaxis=='age':
-                x_age=x
-                x=[]
-            for k in range(1,len(yields_evol)):
-                if sum(yields_evol[k]) ==0:
-                    continue
-                if norm=='no':
-                    y.append(yields_evol[k][iso_idx]/sum(yields_evol[k]))
-                if norm=='ini':
-                    y.append(yields_evol[k][iso_idx]/sum(yields_evol[k])/yields_evol[0][iso_idx])
-
-                x.append(x_age[k])
-            plt.ylabel('X('+yaxis+')')
-            self.y=y
-            plt.yscale('log')
-        elif 'Z' == yaxis:
-            y=self.history.metallicity
-            plt.ylabel('ISM metallicity')
-            plt.yscale('log')
-        elif yaxis in self.history.elements:
-            if source == 'all':
-                yields_evol=self.history.ism_elem_yield
-            elif source =='agb':
-                yields_evol=self.history.ism_elem_yield_agb
-            elif source == 'sn1a':
-                yields_evol=self.history.ism_elem_yield_1a
-            elif source == 'massive':
-                yields_evol=self.history.ism_elem_yield_massive
-            iso_idx=self.history.elements.index(yaxis)
-            y=[]
-            #change of xaxis array 'age 'due to continue statement below
-            if xaxis=='age':
-                x_age=x
-                x=[]
-            for k in range(1,len(yields_evol)):
-                if sum(yields_evol[k]) ==0:
-                    continue
-                if norm=='no':
-                    y.append(yields_evol[k][iso_idx]/sum(yields_evol[k]))
-                if norm=='ini':
-                    y.append(yields_evol[k][iso_idx]/sum(yields_evol[k])/yields_evol[0][iso_idx])
-
-                x.append(x_age[k])
-            #plt.yscale('log')
-            plt.ylabel('X('+yaxis+')')
-            self.y=y
-            plt.yscale('log')
-        #To prevent 0 +log scale
-        if 'age' == xaxis:
-                x=x[1:]
-                y=y[1:]
-		plt.xlim(self.history.dt,self.history.tend)
-        plt.plot(x,y,label=label,linestyle=shape,marker=marker,color=color,markevery=markevery)
-        plt.legend()
-        ax=plt.gca()
-        self.__fig_standard(ax=ax,fontsize=fontsize,labelsize=labelsize,rspace=rspace, bspace=bspace,legend_fontsize=legend_fontsize)
-	self.save_data(header=[xaxis,yaxis],data=[x,y])
-
-    def plot_spectro(self,fig=3,xaxis='age',yaxis='[Fe/H]',source='all',label='',shape='-',marker='o',color='k',markevery=100,show_data=False,show_sculptor=False,show_legend=True,return_x_y=False,sub_plot=False,linewidth=3,sub=1,plot_data=False,fsize=[10,4.5],fontsize=14,rspace=0.6,bspace=0.15,labelsize=15,legend_fontsize=14,only_one_iso=False,solar_ab=''):
-        '''
-        Plots elements in spectroscopic notation:
-
-        Parameters
-        ----------
-
-	xaxis : string
-            Elements spectroscopic notation e.g. [Fe/H]
-	    if 'age': time evolution in years
-	yaxis : string
-                Elements in spectroscopic notation, e.g. [C/Fe]
-	source : string
-                If yields come from
-		all sources use 'all' (include
-		AGB+SN1a, massive stars.)
-
-		If yields come from distinctive source:
-		only agb stars use 'agb', only
-		SN1a ('SN1a'), or only massive stars
-		('massive')
-
-        label : string
-             figure label
-        marker : string
-             figure marker
-        shape : string
-             line style 
-        color : string
-             color of line
-        fig : string,float
-             to name the plot figure       
-
-        Examples
-        ----------
-	>>> plt.plot_spectro('[Fe/H]','[C/Fe]')
-
-        '''
-
-        #Error message if there is the "subplot" has not been provided
-        if sub_plot and sub == 1:
-            print '!! Error - You need to use the \'sub\' parameter and provide the frame for the plot !!'
-            return
-
-        #Operations associated with plot visual aspects
-        if not return_x_y and not sub_plot:
-
-            if len(label)<1:
-                    label=yaxis
-
-            shape,marker,color=self.__msc(source,shape,marker,color)
-
-        #Set the figure output
-        if sub_plot:
-            plt_ps = sub
-        else:
-            plt_ps = plt
-
-
-        #Access solar abundance
-        if len(solar_ab) > 0:
-            iniabu=ry.iniabu(global_path+solar_ab)
-        else:
-            iniabu=ry.iniabu(global_path+'yield_tables/iniabu/iniab2.0E-02GN93.ppn')
-
-        x_ini_iso=iniabu.iso_abundance(self.history.isotopes)
-        elements,x_ini=self._iso_abu_to_elem(x_ini_iso)
-        #to test the different contributions
-        if source == 'all':
-            yields_evol=self.history.ism_elem_yield
-        elif source =='agb':
-            yields_evol=self.history.ism_elem_yield_agb
-        elif source == 'sn1a':
-            yields_evol=self.history.ism_elem_yield_1a
-        elif source == 'massive':
-            yields_evol=self.history.ism_elem_yield_massive
-
-        #Operations associated with plot visual aspects
-        if not return_x_y and not sub_plot:
-            plt.figure(fig, figsize=(fsize[0],fsize[1]))
-
-
-        if 'age' == xaxis:
-            x=self.history.age
-
-            #Operations associated with plot visual aspects
-            if not return_x_y and not sub_plot:
-                #plt.xscale('log')
-                plt.xlabel('Age [yr]')
-
-            self.x=x
-        else:
-
-
-            xaxis_elem1=xaxis.split('/')[0][1:]
-            xaxis_elem2=xaxis.split('/')[1][:-1]
-
-            #print xaxis_elem1, xaxis_elem2
-
-            #X-axis ini values
-            x_elem1_ini=x_ini[elements.index(xaxis_elem1)]
-            x_elem2_ini=x_ini[elements.index(xaxis_elem2)]
-
-            #X-axis gce values
-            elem_idx1=self.history.elements.index(xaxis_elem1)
-            elem_idx2=self.history.elements.index(xaxis_elem2)
-
-            # Take only 56-Fe is needed
-            if only_one_iso:
-                if yaxis_elem1 == 'Fe':
-                    yields_evol = self.history.ism_iso_yield
-                    elem_idx1 = self.history.isotopes.index('Fe-56')
-
-            x=[]
-            for k in range(0,len(yields_evol)):
-                if sum(yields_evol[k]) ==0:
-                    continue
-                #in case no contribution during timestep
-                x1=yields_evol[k][elem_idx1]/sum(yields_evol[k])
-                x2=yields_evol[k][elem_idx2]/sum(yields_evol[k])
-                if x1 <= 0.0 or x2 <= 0.0:
-                    spec = -30.0
-                else:
-                    spec=np.log10(x1/x2) - np.log10(x_elem1_ini/x_elem2_ini)
-                x.append(spec)
-            #Operations associated with plot visual aspects
-            if not return_x_y and not sub_plot:
-                plt.xlabel(xaxis)
-            self.x=x
-
-
-        yaxis_elem1=yaxis.split('/')[0][1:]
-        yaxis_elem2=yaxis.split('/')[1][:-1]
-
-        #y=axis ini values
-        x_elem1_ini=x_ini[elements.index(yaxis_elem1)]
-        x_elem2_ini=x_ini[elements.index(yaxis_elem2)]
-        #print 'Fe_sol = ',x_elem1_ini,' , H_sol = ',x_elem2_ini
-
-        #Y-axis gce values
-        elem_idx1=self.history.elements.index(yaxis_elem1)
-        elem_idx2=self.history.elements.index(yaxis_elem2)
-
-        # Take only 56-Fe is needed
-        if only_one_iso:
-            if yaxis_elem1 == 'Fe':
-                yields_evol = self.history.ism_iso_yield
-                elem_idx1 = self.history.isotopes.index('Fe-56')
-            elif yaxis_elem2 == 'Fe':
-                yields_evol = self.history.ism_iso_yield
-                elem_idx2 = self.history.isotopes.index('Fe-56')
-
-        y=[]
-        if xaxis=='age':
-            x_age=x
-            x=[]
-        for k in range(0,len(yields_evol)):
-            if sum(yields_evol[k]) ==0:
-                continue
-            #in case no contribution during timestep
-            x1=yields_evol[k][elem_idx1]/sum(yields_evol[k])
-            x2=yields_evol[k][elem_idx2]/sum(yields_evol[k])
-            if x1 <= 0.0 or x2 <= 0.0:
-                spec = -30.0
-            else:
-                #print 'Fe_sim = ',x1, ' , H_sim = ',x2
-                spec=np.log10(x1/x2) - np.log10(x_elem1_ini/x_elem2_ini)
-            y.append(spec)
-            if xaxis=='age':
-                x.append(x_age[k])
-        if len(y)==0:
-            print 'Y values all zero'
-        #Operations associated with plot visual aspects
-        if not return_x_y and not sub_plot:
-            plt.ylabel(yaxis)
-        self.y=y
-	#To prevent 0 +log scale
-	if 'age' == xaxis:
-		x=x[1:]
-		y=y[1:]
-                #Operations associated with plot visual aspects
-                if not return_x_y and not sub_plot:
-                    plt.xlim(self.history.dt,self.history.tend)
-
-        #Remove values when SFR is zero, that is when no element is locked into stars
-        i_rem = len(y) - 1
-        while self.history.sfr_abs[i_rem] == 0.0:
-            del y[-1]
-            del x[-1]
-            i_rem -= 1
-
-        if not plot_data:
-
-          # Filtrate bad value
-          x_temp = []
-          y_temp = []
-          for i_temp in range(0,len(x)):
-              if np.isfinite(x[i_temp]) and np.isfinite(y[i_temp])\
-                 and x[i_temp] > -20.0 and y[i_temp] > -20.0:
-                  x_temp.append(x[i_temp])
-                  y_temp.append(y[i_temp])
-          x = x_temp
-          y = y_temp
-
-          #If this function is supposed to return the x, y arrays only ...
-          if return_x_y:
-
-            return x, y
-
-          #If this is a sub-figure managed by an external module
-          elif sub_plot:
-
-            if self.galaxy == 'none':
-                if show_legend:
-                    sub.plot(x,y,linestyle=shape,label=label,marker=marker,color=color,markevery=markevery,linewidth=linewidth)
-                else:
-                    sub.plot(x,y,linestyle=shape,marker=marker,color=color,markevery=markevery,linewidth=linewidth)
-            else:
-                if show_legend:
-                    sub.plot(x,y,linestyle=shape,label=label,marker=marker,color=color,markevery=markevery,linewidth=linewidth)
-                else:
-                    sub.plot(x,y,linestyle=shape,marker=marker,color=color,markevery=markevery,linewidth=linewidth) 
-
-          #If this function is supposed to plot ...
-          else:
-
-            #Plot a thicker line for specific galaxies, since they must be visible with all the obs. data
-            if self.galaxy == 'none':
-                if show_legend:
-                    plt.plot(x,y,linestyle=shape,label=label,marker=marker,color=color,markevery=markevery,linewidth=linewidth)
-                else:
-                    plt.plot(x,y,linestyle=shape,marker=marker,color=color,markevery=markevery,linewidth=linewidth)
-            else:
-                if show_legend:
-                    plt.plot(x,y,linestyle=shape,label=label,marker=marker,color=color,markevery=markevery,linewidth=linewidth)
-                else:
-                    plt.plot(x,y,linestyle=shape,marker=marker,color=color,markevery=markevery,linewidth=linewidth)
-
-            #plt.plot([1.93,2.123123,3.23421321321],[4.123123132,5.214124142,6.11111],linestyle='--')
-            #plt.plot([1.93,2.123123,3.23421321321],[4.123123132,5.214124142,6.11111],linestyle='--')
-	    if len(label)>0:
-	            plt.legend()
-            ax=plt.gca()
-            self.__fig_standard(ax=ax,fontsize=fontsize,labelsize=labelsize,rspace=rspace, bspace=bspace,legend_fontsize=legend_fontsize)
-	    self.save_data(header=[xaxis,yaxis],data=[x,y])
-
-
-    def plot_totmasses(self,fig=4,mass='gas',source='all',norm='no',label='',shape='',marker='',color='',markevery=20,log=True,fsize=[10,4.5],fontsize=14,rspace=0.6,bspace=0.15,labelsize=15,legend_fontsize=14):
-        '''
-	Plots either gas or star mass in fraction of total mass
-	vs time.
-        
-        Parameters
-        ----------
-
-        mass : string
-            either 'gas' for ISM gas mass
-            or 'stars' for gas locked away in stars (totalgas - ISM gas)
-
-        norm : string
-            normalization, either 'no' for no normalization (total gass mass in solar masses),
-
-	    for normalization to the initial gas mass (mgal) with 'ini',
-
-	    for normalization to the current total gas mass 'current'.
-	    The latter case makes sense when comparing different
-	    sources (see below)
-
-        source : string
-            specificies if yields come from
-	    all sources ('all'), including
-	    AGB+SN1a, massive stars. Or from
-	    distinctive sources:
-	    only agb stars ('agb'), only
-	    SN1a ('sn1a'), or only massive stars
-	    ('massive')
-	log : boolean
-            if true plot logarithmic y axis
-        label : string
-             figure label
-        marker : string
-             figure marker
-        shape : string
-             line style 
-        color : string
-             color of line
-        fig : string,float
-             to name the plot figure       
-
-        Examples
-        ----------
-        >>> s.plot_totmasses()
-
-
-        '''
-
-        import matplotlib.pyplot as plt
-
-        #if len(label)<1:
-        #        label=mass+', '+source
-
-
-        plt.figure(fig, figsize=(fsize[0],fsize[1]))
-
-        #Assume isotope input
-
-	xaxis='age'
-	if source =='all':
-		if len(label)==0:
-			label='All'
-	if source == 'agb':
-		if len(label)==0:
-			label='AGB'
-	if source =='massive':
-		if len(label)==0:
-			label='Massive'
-	if source =='sn1a':
-		if len(label)==0:
-			label='SNIa'
-	
-	shape,marker,color=self.__msc(source,shape,marker,color)
-
-	if 'age' == xaxis:
-	    x_all=self.history.age#[1:]
-	    plt.xscale('log')
-	    plt.xlabel('log-scaled '+xaxis+' [yrs]')
-	    #self.x=x
-
-	gas_mass=self.history.gas_mass
-
-	#to test the different contributions
-	if source == 'all':
-            gas_evol=self.history.gas_mass
-        else:
-            if source =='agb':
-	        yields_evol=self.history.ism_elem_yield_agb
-            elif source == 'sn1a':
-	        yields_evol=self.history.ism_elem_yield_1a
-            elif source == 'massive':
-	        yields_evol=self.history.ism_elem_yield_massive
-            gas_evol=[]
-            for k in range(len(yields_evol)):
-	        gas_evol.append(sum(yields_evol[k]))
-
-	ism_gasm=[]
-	star_m=[]
-	x=[]
-        #To prevent 0 +log scale
-        if 'age' == xaxis:
-                x_all=x_all[1:]
-                gas_evol=gas_evol[1:]
-		gas_mass=gas_mass[1:]
-	for k in range(0,len(gas_evol)):
-	    if (gas_evol[k]==0) or (gas_mass[k]==0):
-		continue
-	    x.append(x_all[k])
-            if norm=='ini':
-	        ism_gasm.append(gas_evol[k]/self.history.mgal)
-	        star_m.append((self.history.mgal-gas_evol[k])/self.history.mgal)
-            if norm == 'current':
-	        if not self.history.gas_mass[k] ==0.:
-	              ism_gasm.append(gas_evol[k]/gas_mass[k])
-	              star_m.append((self.history.mgal-gas_evol[k])/gas_mass[k])
-	    #else:
-	     #   ism_gasm.append(0.)
-	      #  star_m.append(0.)
-            elif norm == 'no':
-	        ism_gasm.append(gas_evol[k])
-	        star_m.append(self.history.mgal-gas_evol[k])
-        if mass == 'gas':
-	    y=ism_gasm
-        if mass == 'stars':
-	    y=star_m
-	plt.plot(x,y,linestyle=shape,marker=marker,markevery=markevery,color=color,label=label)	
-        if len(label)>0:
-		plt.legend()
-        if norm=='current':
-            plt.ylim(0,1.2)
-        if not norm=='no':
-            if mass=='gas':
-                plt.ylabel('mass fraction')
-                plt.title('Gas mass as a fraction of total gas mass')
-            else:
-                plt.ylabel('mass fraction')
-                plt.title('Star mass as a fraction of total star mass')
-        else:
-            if mass=='gas':
-                plt.ylabel('ISM gas mass [Msun]')
-            else:
-                plt.ylabel('mass locked in stars [Msun]')
-
-            if mass=='gas':
-                plt.ylabel('ISM gas mass [Msun]')
-            else:
-                plt.ylabel('Mass locked in stars [Msun]')
-
-        if log==True:
-            plt.yscale('log')
-	    if not norm=='no':
-            	plt.ylim(1e-4,1.2)
-        ax=plt.gca()
-        self.__fig_standard(ax=ax,fontsize=fontsize,labelsize=labelsize,rspace=rspace, bspace=bspace,legend_fontsize=legend_fontsize)
-	plt.xlim(self.history.dt,self.history.tend)
-	self.save_data(header=['age','mass'],data=[x,y])
-
-
-    def plot_sn_distr(self,fig=5,rate=True,rate_only='',xaxis='time',fraction=False,label1='SNIa',label2='SN2',shape1=':',shape2='--',marker1='o',marker2='s',color1='k',color2='b',markevery=20,fsize=[10,4.5],fontsize=14,rspace=0.6,bspace=0.15,labelsize=15,legend_fontsize=14):
-
-        '''
-	Plots the SN1a distribution:
-	The evolution of the number of SN1a and SN2
-	#Add numbers/dt numbers/year...
-
-        Parameters
-        ----------
-	rate : boolean
-            if true, calculate rate [1/century] 
-            else calculate numbers
-	fraction ; boolean
-	    if true, ignorate rate and calculate number fraction of SNIa per WD
-        rate_only : string
-	    if empty string, plot both rates (default)
-
-            if 'sn1a', plot only SN1a rate
-
-            if 'sn2', plot only SN2 rate
-	xaxis: string
-	    if 'time' : time evolution
-	    if 'redshift': experimental! use with caution; redshift evolution
-        label : string
-             figure label
-        marker : string
-             figure marker
-        shape : string
-             line style 
-        color : string
-             color of line
-        fig : string,float
-             to name the plot figure       
-
-        Examples
-        ----------
-        >>> s.plot_sn_distr()
-        '''
-	#For Wiersma09
-	Hubble_0=73.
-	Omega_lambda=0.762
-	Omega_m=0.238
-
-        figure=plt.figure(fig, figsize=(fsize[0],fsize[1]))
-        age=self.history.age
-        sn1anumbers=self.history.sn1a_numbers#[:-1]
-        sn2numbers=self.history.sn2_numbers
-	if xaxis=='redshift':
-		print 'this features is not tested yet.'
-		return 0
-		age,idx=self.__time_to_z(age,Hubble_0,Omega_lambda,Omega_m)
-		age=[0]+age
-		plt.xlabel('Redshift z')
-		timesteps=self.history.timesteps[idx-1:]
-		sn2numbers=sn2numbers[idx:]
-		sn1anumbers=sn1anumbers[idx:]
-	else:
-		plt.xlabel('Log-scaled age [yrs]')
-		#plt.xscale('log')
-        if rate and not fraction:
-	    if xaxis=='redshift':
-			sn1a_rate=np.array(sn1anumbers)/ (np.array(timesteps)/100.)
-			sn2_rate=np.array(sn2numbers)/ (np.array(timesteps)/100.)	
-	    else:
-            		sn1a_rate=np.array(sn1anumbers[1:])/ (np.array(self.history.timesteps)/100.)
-            		sn2_rate=np.array(sn2numbers[1:])/ (np.array(self.history.timesteps)/100.)
-	    sn1a_rate1=[]
-	    sn2_rate1=[]
-	    age=age[1:]
-	    age_sn1a=[] #age[1:]
-	    age_sn2=[]
-	    #correct sn1a rate
-	    for k in range(len(sn1a_rate)):
-		if sn1a_rate[k]>0:
-			sn1a_rate1.append(sn1a_rate[k])
-			age_sn1a.append(age[k])
-			
-	    for k in range(len(sn2_rate)):
-		if sn2_rate[k]>0:
-			sn2_rate1.append(sn2_rate[k])
-			age_sn2.append(age[k])
-	
-
-            if len(rate_only)==0:
-		    x=[age_sn2,age_sn1a]
-		    y=[sn1a_rate1,sn2_rate1]
-		    plt.plot(age_sn1a,sn1a_rate1,linestyle=shape1,color=color1,label=label1,marker=marker1,markevery=markevery)
-		    plt.plot(age_sn2,sn2_rate1,linestyle=shape2,color=color2,label=label2,marker=marker2,markevery=markevery)
-            	    plt.ylabel('SN rate [century$^{-1}$]')
-	    if rate_only=='sn1a':
-		    x=age_sn1a
-		    y= sn1a_rate1
-		    plt.plot(age_sn1a,sn1a_rate1,linestyle=shape1,color=color1,label=label1,marker=marker1,markevery=markevery)
-		    plt.ylabel('SNIa rate [century$^{-1}$]')
-	    if rate_only=='sn2':
-		    x=age_sn2
-		    y=sn2_rate
-                    plt.plot(age_sn2,sn2_rate1,linestyle=shape2,color=color2,label=label2,marker=marker2,markevery=markevery)
-		    plt.ylabel('SN2 rate [century$^{-1}$]')
-
-
-        else:
-            #if xaxis=='redshift':
-		
-                        #sn1_numbers=np.array(sn1anumbers)/ (np.array(timesteps)/100.)
-                        #sn2_numbers=np.array(sn2numbers)/ (np.array(timesteps)/100.)
-            #True: #else:
-                        #sn1a_rate=np.array(sn1anumbers[1:])/ (np.array(self.history.timesteps)/100.)
-                        #sn2_rate=np.array(sn2numbers[1:])/ (np.array(self.history.timesteps)/100.)
-
-	    sn1a_numbers=sn1anumbers[1:]
-	    sn2_numbers=sn2numbers[1:]	
-            sn1a_numbers1=[]
-	    sn2_numbers1=[]
-            age_sn1a=[]
-	    age_sn2=[]
-            age=age[1:]
-            for k in range(len(sn1a_numbers)):
-                if sn1a_numbers[k]>0:
-                        sn1a_numbers1.append(sn1a_numbers[k])
-                        age_sn1a.append(age[k])
-	    for k in range(len(sn2_numbers)):
-                if sn2_numbers[k]>0:
-                        sn2_numbers1.append(sn2_numbers[k])
-                        age_sn2.append(age[k])
-
-	    if fraction:
-		age=self.history.age
-		ratio=[]
-		age1=[]
-		for k in range(len(self.wd_sn1a_range1)):
-		    if self.wd_sn1a_range1[k]>0:
-			ratio.append(self.history.sn1a_numbers[1:][k]/self.wd_sn1a_range1[k])
-			age1.append(age[k])
-		plt.plot(age1,ratio)
-		plt.yscale('log')
-		#plt.xscale('log')
-		plt.ylabel('Number of SNIa going off per WD born')
-		label='SNIafractionperWD';label='sn1a '+label
-		x=age1
-		y=ratio
-		self.save_data(header=['age',label],data=[x,y])		
-		return 
-	    else:
-		    if len(rate_only)==0:
-			    x=[age_sn1a,age_sn2]
-			    y=[sn1a_numbers,sn2_numbers]
-			    plt.plot(age_sn1a,sn1a_numbers1,linestyle=shape1,color=color1,label=label1,marker=marker1,markevery=markevery)
-			    plt.plot(age_sn2,sn2_numbers1,linestyle=shape2,color=color2,label=label2,marker=marker2,markevery=markevery)
-			    plt.ylabel('SN numbers')
-		    if rate_only=='sn1a':
-			    x= age1
-			    y= sn1anumbers1
-			    plt.plot(age_sn1a,sn1a_numbers1,linestyle=shape1,color=color1,label=label1,marker=marker1,markevery=markever)
-			    plt.ylabel('SN numbers')
-		    if rate_only=='sn2':
-			    x= age[1:]
-			    y= sn2numbers[1:]
-			    plt.plot(age_sn2,sn2_numbers1,linestyle=shape2,color=color2,label=label2,marker=marker2,markevery=markevery)
-			    plt.ylabel('SN numbers')
-
-	plt.legend(loc=1)
-	plt.yscale('log')
-	ax=plt.gca()
-        self.__fig_standard(ax=ax,fontsize=fontsize,labelsize=labelsize,rspace=rspace, bspace=bspace,legend_fontsize=legend_fontsize)
-	if rate:
-		label='rate'
-	else:
-		label='number'	
-	if len(rate_only)==0:
-		if rate:
-			label='rate'
-		else:
-			label='number'
-		self.save_data(header=['age','SNIa '+label,'age','CCSN '+label],data=[x[0],y[0],x[1],y[1]])
-	else:
-		if rate_only=='sn1a':
-			label='sn1a '+label
-		else:
-			label='ccsn '+label
-                self.save_data(header=['age',label],data=[x,y])
-
-
     def save_data(self,header=[],data=[],filename='plot_data.txt'):
 	'''
 		Writes data into a text file. data entries
@@ -2710,128 +1823,6 @@ class omega( chem_evol ):
 	f1.write(out)
 	f1.close()
 
-
-
-    ##############################################
-    #          Plot Star Formation Rate          #
-    ##############################################
-    def plot_star_formation_rate(self,fig=6,fraction=False,source='all',marker='',shape='',color='',label='',abs_unit=True,fsize=[10,4.5],fontsize=14,rspace=0.6,bspace=0.15,labelsize=15,legend_fontsize=14):
-	'''
-	Plots the star formation rate over time.
-	Shows fraction of ISM mass which transforms into stars.
-
-        Parameters
-        ----------
-
-	fraction : boolean
-		if true: fraction of ISM which transforms into stars;
-		else: mass of ISM which goes into stars
-
-	source : string
-		either 'all' for total star formation rate; 'agb' for AGB and 'massive' for massive stars
-
-	marker : string
-		marker type
-	shape : string
-		line shape type
-	fig: figure id
-	
-        Examples
-        ----------
-        >>> s.star_formation_rate()
-	
-	'''
-
-	if (len(marker)==0 and len(shape)==0) and len(color)==0:
-		shape,marker,color=self.__msc(source,shape,marker,color)
-	plt.figure(fig, figsize=(fsize[0],fsize[1]))
-	#maybe a histogram for display the SFR?
-	if False:
-		age=self.history.age
-		#age=[0.1]+self.history.age[1:-1]
-		sfr=self.history.sfr
-		plt.xlabel('Log-scaled age [yrs]')
-		#plt.xscale('log')
-		mean_age=[]
-		color='r'
-		label='aa'
-		age_bdy=[]
-		for k in range(len(age)-1):
-			mean_age.append(1) #age[k+1]-age[k])
-		for k in range(len(age)):
-			age_bdy.append(age[k])	
-		print 'age',len(mean_age)
-		print 'weights',len(sfr)
-		print 'bdy',len(age_bdy)
-		print sfr[:5]
-	        p1 =plt.hist(mean_age, bins=age_bdy,weights=sfr,facecolor=color,color=color,alpha=0.5,label=label)
-
-        #Plot the SFR in [Mo yr^-1]
-        if abs_unit:
-
-            #Define visual aspects
-            shape,marker,color=self.__msc(source,shape,marker,color)
-
-            #Recover the SFR by dividing the mass locked into stars, at
-            #each timestep, by the duration of the timestep
-            age=self.history.age
-            #sfr_plot = self.history.m_locked / self.history.timesteps
-            sfr_plot = self.history.sfr_abs
-            
-            #Label and display axis
-            plt.xlabel('Age [yrs]')
-            plt.ylabel('SFR [Mo/yr]')	
-
-            #Plot
-            plt.plot(age,sfr_plot,label=label,marker=marker,color=color,linestyle=shape)
-
-            #self.save_data(header=['age','SFR'],data=[age,sfr_plot])
-
-        #Plot the mass fraction of gas available converted into stars
-	elif fraction:
-		sfr=self.history.sfr
-                if source=='all':
-                        label='all'
-                elif source=='agb':
-			sfr=np.array(sfr)*np.array(self.history.m_locked_agb)/np.array(self.history.m_locked)
-                        label='AGB'
-                elif source=='massive':
-                        masslocked=self.history.m_locked_massive
-			sfr=np.array(sfr)*np.array(self.history.m_locked_massive)/np.array(self.history.m_locked)
-                        label='Massive'
-		age=self.history.age[:-1]
-		print len(age),len(sfr)
-		plt.xlabel('Age [yrs]')
-		plt.plot(age,sfr,label=label,marker=marker,linestyle=shape)
-		plt.ylabel('Fraction of current gas mass into stars')	
-        	self.save_data(header=['age','SFR'],data=[age,sfr])
-
-        #Plot the mass converted into stars
-	else:
-		if source=='all':
-			masslocked=self.history.m_locked
-			label='all'
-		elif source=='agb':
-			masslocked=self.history.m_locked_agb
-			label='AGB'
-		elif source=='massive':
-			masslocked=self.history.m_locked_massive
-			label='Massive'
-		age=self.history.age[1:]
-		plt.plot(age,masslocked,marker=marker,linestyle=shape,label=label)
-		plt.xlabel('Age [yrs]')
-		plt.ylabel('ISM Mass transformed into stars')
-		plt.xscale('log');plt.yscale('log')
-		plt.legend()
-
-        ax=plt.gca()
-        self.__fig_standard(ax=ax,fontsize=fontsize,labelsize=labelsize,rspace=rspace, bspace=bspace,legend_fontsize=legend_fontsize)
-
-	#print 'Total mass transformed in stars, total mass transformed in AGBs, total mass transformed in massive stars:'
-	#print sum(self.history.m_locked),sum(self.history.m_locked_agb),sum(self.history.m_locked_massive)
-
-
-
     def __time_to_z(self,time,Hubble_0,Omega_lambda,Omega_m):
 	'''
 	Time to redshift conversion
@@ -2848,1032 +1839,6 @@ class omega( chem_evol ):
 				index=k
 				firstidx=False	
 	return tz.t_to_z(time_new,Hubble_0,Omega_lambda,Omega_m),index
-
-    def __msc(self,source,shape,marker,color):
-
-	'''
-	Function checks if either of shape,color,marker
-	is set. If not then assign in each case
-	a unique property related to the source ('agb','massive'..)
-	to the variable and returns all three
-	'''
-
-	if source=='all':
-		shape1='-'
-		marker1='o'
-		color1='k'
-	elif source=='agb':
-		shape1='--'
-		marker1='s'
-		color1='r'
-	elif source=='massive':
-		shape1='-.'
-		marker1='D'
-		color1='b'
-	elif source =='sn1a':
-		shape1=':'
-		marker1='x'
-		color1='g'
-	
-	if len(shape)==0:
-		shape=shape1
-	if len(marker)==0:
-		marker=marker1
-	if len(color)==0:
-		color=color1
-
-	return shape,marker,color
-
-
-    ##############################################
-    #                Fig Standard                #
-    ##############################################
-    def __fig_standard(self,ax,fontsize=8,labelsize=8,lwtickboth=[6,2],lwtickmajor=[10,3],markersize=8,rspace=0.6,bspace=0.15, legend_fontsize=14):
-
-        '''
-	Internal function in order to get standardized figure font sizes.
-	It is used in the plotting functions.
-        '''
-
-        plt.legend(loc=2,prop={'size':legend_fontsize})
-        plt.rcParams.update({'font.size': fontsize})
-        ax.yaxis.label.set_size(labelsize)
-        ax.xaxis.label.set_size(labelsize)
-	#ax.xaxis.set_tick_params(width=2)
-	#ax.yaxis.set_tick_params(width=2)		
-	ax.tick_params(length=lwtickboth[0],width=lwtickboth[1],which='both')
-	ax.tick_params(length=lwtickmajor[0],width=lwtickmajor[1],which='major')
-	#Add that line below at some point
-	#ax.xaxis.set_tick_params(width=2)
-	#ax.yaxis.set_tick_params(width=2)
-	if len(ax.lines)>0:
-		for h in range(len(ax.lines)):
-			ax.lines[h].set_markersize(markersize)
-	ax.legend(loc='center left', bbox_to_anchor=(1.01, 0.5),markerscale=0.8,fontsize=legend_fontsize)
-
-        plt.subplots_adjust(right=rspace)
-        plt.subplots_adjust(bottom=bspace)
-
-
-    ##############################################
-    #              Plot Mass-Loading             #
-    ##############################################
-    def plot_mass_loading(self,fig=8,marker='',shape='',\
-            color='',label='Mass-loading',fsize=[10,4.5],fontsize=14,rspace=0.6,\
-            bspace=0.15,labelsize=15,legend_fontsize=14):
-
-        '''
-	This function plots the mass-loading factor, which is the ratio
-	between the mass outflow rate and the star formation rate, as a
-	function of time.
-
-        Parameters
-        ----------
-
-        fig : string, float
-	     Figure name.
-	marker : string
-	     Figure marker.
-	shape : string
-	     Line style.
-	color : string
-	     Line color.
-	label : string
-	     Figure label.
-	fsize : 2D float array
-	     Figure dimension/size.
-	fontsize : integer
-	     Font size of the numbers on the X and Y axis.
-	rspace : float
-	     Extra space on the right for the legend.
-	bspace : float
-	     Extra space at the bottom for the Y axis label.
-	labelsize : integer
-	     Font size of the X and Y axis labels.
-	legend_fontsize : integer
- 	       Font size of the legend.
-
-        Examples
-	----------
-
-	>>> o1.plot_mass_loading(label='Mass-Loading Factor')
-
-        '''
-
-        # Define variable only used for plot visual
-        source='all'
-
-        #Plot only if in the open box scenario
-        if self.open_box:
- 
-            #Define visual aspects
-            shape,marker,color=self.__msc(source,shape,marker,color)
-            plt.figure(fig, figsize=(fsize[0],fsize[1]))
-
-            #Display axis labels
-            plt.xlabel('Age [yrs]')
-            plt.ylabel('Mass-loading factor')	
-
-            #Copy data for x and y axis
-            age = self.history.age[:-1]
-            mass_lo = self.history.eta_outflow_t
-
-            #Plot data
-            plt.plot(age,mass_lo,color=color,label=label,\
-                     marker=marker,linestyle=shape)
-
-            #Save plot
-            self.save_data(header=['age','mass-loading'],data=[age,mass_lo])
-
-            ax=plt.gca()
-            self.__fig_standard(ax=ax,fontsize=fontsize,labelsize=labelsize,\
-                rspace=rspace, bspace=bspace,legend_fontsize=legend_fontsize)
-
-        else:
-            print 'Not available with a closed box.'
-
-
-    ##############################################
-    #              Plot Outflow Rate             #
-    ##############################################
-    def plot_outflow_rate(self,fig=9,marker='',shape='',\
-            color='',label='Outflow',fsize=[10,4.5],fontsize=14,rspace=0.6,\
-            bspace=0.15,labelsize=15,legend_fontsize=14):
-
-        '''
-	This function plots the mass outflow rate as a function of time
-        in units of [Mo/yr].
-
-        Parameters
-        ----------
-
-        fig : string, float
-	     Figure name.
-	marker : string
-	     Figure marker.
-	shape : string
-	     Line style.
-	color : string
-	     Line color.
-	label : string
-	     Figure label.
-	fsize : 2D float array
-	     Figure dimension/size.
-	fontsize : integer
-	     Font size of the numbers on the X and Y axis.
-	rspace : float
-	     Extra space on the right for the legend.
-	bspace : float
-	     Extra space at the bottom for the Y axis label.
-	labelsize : integer
-	     Font size of the X and Y axis labels.
-	legend_fontsize : integer
- 	       Font size of the legend.
-
-        Examples
-	----------
-
-	>>> o1.plot_outflow_rate(label='Outflow Rate')
-
-        '''
-
-        # Define variable only used for plot visual
-        source='all'
-
-        #Plot only if in the open box scenario
-        if self.open_box:
-
-            #Define visual aspects
-            shape,marker,color=self.__msc(source,shape,marker,color)
-            plt.figure(fig, figsize=(fsize[0],fsize[1]))
-
-            #Display axis labels
-            plt.xlabel('Age [yrs]')
-            plt.ylabel('Outflow rate [Mo/yr]')	
-
-            #Copy data for x and y axis
-            age = self.history.age[:-1]
-            outflow_plot = []
-            for i_op in range(0,len(age)):
-                outflow_plot.append(self.history.m_outflow_t[i_op] / \
-                    self.history.timesteps[i_op])
-            outflow_plot[0] = 0.0
-
-            #Plot data
-            plt.plot(age,outflow_plot,label=label,marker=marker,\
-                     color=color,linestyle=shape)
-
-            #Save plot
-            self.save_data(header=['age','outflow rate'],data=[age,outflow_plot])
-
-            ax=plt.gca()
-            self.__fig_standard(ax=ax,fontsize=fontsize,labelsize=labelsize,\
-                rspace=rspace, bspace=bspace,legend_fontsize=legend_fontsize)
-
-        else:
-            print 'Not available with a closed box.'
-
-
-    ##############################################
-    #              Plot Inflow Rate              #
-    ##############################################
-    def plot_inflow_rate(self,fig=10,marker='',shape='',color='',\
-            label='Inflow',fsize=[10,4.5],fontsize=14,rspace=0.6,bspace=0.15,\
-            labelsize=15,legend_fontsize=14):
-
-        '''
-	This function plots the mass inflow rate as a function of time
-        in units of [Mo/yr].
-
-        Parameters
-        ----------
-
-        fig : string, float
-	     Figure name.
-	marker : string
-	     Figure marker.
-	shape : string
-	     Line style.
-	color : string
-	     Line color.
-	label : string
-	     Figure label.
-	fsize : 2D float array
-	     Figure dimension/size.
-	fontsize : integer
-	     Font size of the numbers on the X and Y axis.
-	rspace : float
-	     Extra space on the right for the legend.
-	bspace : float
-	     Extra space at the bottom for the Y axis label.
-	labelsize : integer
-	     Font size of the X and Y axis labels.
-	legend_fontsize : integer
- 	       Font size of the legend.
-
-        Examples
-	----------
-
-	>>> o1.plot_inflow_rate(label='Inflow Rate')
-
-        '''
-
-        # Define variable only used for plot visual
-        source='all'
-
-        #Plot only if in the open box scenario
-        if self.open_box:
-
-            #Define visual aspects
-            shape,marker,color=self.__msc(source,shape,marker,color)
-	    plt.figure(fig, figsize=(fsize[0],fsize[1]))
-
-            #Display axis labels
-            plt.xlabel('Age [yrs]')
-            plt.ylabel('Inflow rate [Mo/yr]')	
-
-            #Copy data for x and y axis
-            age = self.history.age[:-1]
-            inflow_plot = []
-            for i_op in range(0,len(age)):
-                inflow_plot.append(self.history.m_inflow_t[i_op] / \
-                    self.history.timesteps[i_op])
-
-            #Plot data
-            plt.plot(age,inflow_plot,label=label,marker=marker,\
-                     color=color,linestyle=shape)
-
-            #Save plot
-            self.save_data(header=['age','inflow rate'],data=[age,inflow_plot])
-
-            ax=plt.gca()
-            self.__fig_standard(ax=ax,fontsize=fontsize,labelsize=labelsize,\
-                rspace=rspace, bspace=bspace,legend_fontsize=legend_fontsize)
-
-        else:
-            print 'Not available with a closed box.'
-
-
-    ##############################################
-    #              Plot Dark Matter              #
-    ##############################################
-    def plot_dark_matter(self,fig=11,marker='',shape='',\
-            color='',label='Dark matter',fsize=[10,4.5],fontsize=14,rspace=0.6, \
-            bspace=0.15,labelsize=15,legend_fontsize=14):
-
-        '''
-	This function plots the dark matter halo mass of the galaxy as a 
-	function of time.
-
-        Parameters
-        ----------
-
-        fig : string, float
-	     Figure name.
-	marker : string
-	     Figure marker.
-	shape : string
-	     Line style.
-	color : string
-	     Line color.
-	label : string
-	     Figure label.
-	fsize : 2D float array
-	     Figure dimension/size.
-	fontsize : integer
-	     Font size of the numbers on the X and Y axis.
-	rspace : float
-	     Extra space on the right for the legend.
-	bspace : float
-	     Extra space at the bottom for the Y axis label.
-	labelsize : integer
-	     Font size of the X and Y axis labels.
-	legend_fontsize : integer
- 	       Font size of the legend.
-
-        Examples
-	----------
-
-	>>> o1.plot_dark_matter(label='Dark Matter')
-
-        '''
-
-        # Define variable only used for plot visual
-        source='all'
-
-        # Plot only if in the open box scenario
-        if self.open_box:
-
-            # Define visual aspects
-            shape,marker,color=self.__msc(source,shape,marker,color)
-	    plt.figure(fig, figsize=(fsize[0],fsize[1]))
-
-            #Display axis labels
-            plt.xlabel('Age [yrs]')
-            plt.ylabel('Dark matter halo mass [Mo]')	
-
-            #Copy data for x and y axis
-            age = self.history.age
-            DM_plot = self.history.m_DM_t
-
-            #Plot data
-            plt.plot(age,DM_plot,label=label,marker=marker,\
-                     color=color,linestyle=shape)
-
-            #Save plot
-            self.save_data(header=['age','dark matter halo mass'],\
-                           data=[age,DM_plot])
-
-            ax=plt.gca()
-            self.__fig_standard(ax=ax,fontsize=fontsize,labelsize=labelsize,\
-                rspace=rspace, bspace=bspace,legend_fontsize=legend_fontsize)
-
-        else:
-            print 'Not available with a closed box.'
-
-
-    ##############################################
-    #             Plot SF Timescale              #
-    ##############################################
-    def plot_sf_timescale(self,fig=15,marker='',shape='',color='',\
-            label='',fsize=[10,4.5],fontsize=14,rspace=0.6,bspace=0.15,\
-            labelsize=15,legend_fontsize=14):
-
-        '''
-	This function plots the star formation timescale as a function of time
-        in units of [yr].
-
-        Parameters
-        ----------
-
-        fig : string, float
-	     Figure name.
-	marker : string
-	     Figure marker.
-	shape : string
-	     Line style.
-	color : string
-	     Line color.
-	label : string
-	     Figure label.
-	fsize : 2D float array
-	     Figure dimension/size.
-	fontsize : integer
-	     Font size of the numbers on the X and Y axis.
-	rspace : float
-	     Extra space on the right for the legend.
-	bspace : float
-	     Extra space at the bottom for the Y axis label.
-	labelsize : integer
-	     Font size of the X and Y axis labels.
-	legend_fontsize : integer
- 	       Font size of the legend.
-
-        Examples
-	----------
-
-	>>> o1.plot_sf_timescale(label='Star Formation Timescale')
-
-        '''
-
-        # Define variable only used for plot visual
-        source='all'
-
-        #Plot only if in the open box scenario
-        if self.open_box:
-
-            #Define visual aspects
-            shape,marker,color=self.__msc(source,shape,marker,color)
-	    plt.figure(fig, figsize=(fsize[0],fsize[1]))
-
-            #Display axis labels
-            plt.xlabel('Age [yrs]')
-            plt.ylabel('Star formation timescale [yr]')	
-
-            #print len(self.history.age)
-            #print len(self.history.t_SF_t)
-
-            #Copy data for x and y axis
-            age = self.history.age
-            t_SF_plot = self.history.t_SF_t
-
-            #Plot data
-            plt.plot(age,t_SF_plot,label=label,marker=marker,\
-                     color=color,linestyle=shape)
-
-            #Save plot
-            self.save_data(header=['age','star formation timescale'],\
-                                   data=[age,t_SF_plot])
-
-            ax=plt.gca()
-            self.__fig_standard(ax=ax,fontsize=fontsize,labelsize=labelsize,\
-                rspace=rspace, bspace=bspace,legend_fontsize=legend_fontsize)
-
-        else:
-            print 'Not available with a closed box.'
-
-
-    ##############################################
-    #               Plot Redshift                #
-    ##############################################
-    def plot_redshift(self,fig=16,marker='',shape='',\
-            color='',label='Redshift',fsize=[10,4.5],fontsize=14,rspace=0.6,\
-            bspace=0.15,labelsize=15,legend_fontsize=14):
-
-        '''
-	This function plots the redshift a function of time.
-
-        Parameters
-        ----------
-
-        fig : string, float
-	     Figure name.
-	marker : string
-	     Figure marker.
-	shape : string
-	     Line style.
-	color : string
-	     Line color.
-	label : string
-	     Figure label.
-	fsize : 2D float array
-	     Figure dimension/size.
-	fontsize : integer
-	     Font size of the numbers on the X and Y axis.
-	rspace : float
-	     Extra space on the right for the legend.
-	bspace : float
-	     Extra space at the bottom for the Y axis label.
-	labelsize : integer
-	     Font size of the X and Y axis labels.
-	legend_fontsize : integer
- 	       Font size of the legend.
-
-        Examples
-	----------
-
-	>>> o1.plot_redshift(label='Redshift')
-
-        '''
-
-        # Define variable only used for plot visual
-        source='all'
-
-        #Plot only if in the open box scenario
-        if self.open_box:
-
-            #Define visual aspects
-            shape,marker,color=self.__msc(source,shape,marker,color)
-            plt.figure(fig, figsize=(fsize[0],fsize[1]))
-
-            #Display axis labels
-            plt.xlabel('Age [yrs]')
-            plt.ylabel('Redshift')	
-
-            #Copy data for x and y axis
-            age = self.history.age
-            redshift_plot = self.history.redshift_t
-
-            #Plot data
-            plt.plot(age,redshift_plot,label=label,marker=marker,\
-                     color=color,linestyle=shape)
-
-            #Save plot
-            self.save_data(header=['age','redshift'],data=[age,redshift_plot])
-
-            ax=plt.gca()
-            self.__fig_standard(ax=ax,fontsize=fontsize,labelsize=labelsize,\
-                rspace=rspace, bspace=bspace,legend_fontsize=legend_fontsize)
-
-        else:
-            print 'Not available with a closed box.'
-
-
-    ###################################################
-    #                  Plot Iso Ratio                 #
-    ###################################################
-    def plot_iso_ratio(self,return_x_y=False,
-        xaxis='age',yaxis='C-12/C-13',\
-        solar_ab='yield_tables/iniabu/iniab2.0E-02GN93.ppn',\
-        solar_iso='solar_normalization/Asplund_et_al_2009_iso.txt',\
-        fig=18,source='all',marker='',shape='',\
-        color='',label='',fsize=[10,4.5],fontsize=14,rspace=0.6,\
-        bspace=0.15,labelsize=15,legend_fontsize=14):
-
-        '''
-	This function plots the evolution of an isotopic ratio as a function
-	of time, as a function of an abundance ratio, or as a function of
-	another isotopic ratio.  Isotopic ratios are given in the
-	delta notation and abundances ratios are given in spectroscopic notation.
-
-        Parameters
-        ----------
-
-	return_x_y : boolean
-	     If False, show the plot.  If True, return two arrays containing
-	     the X and Y axis data, respectively.
-	xaxis : string
-	     X axis, either 'age', an abundance ratio such as '[Fe/H]', or an
-	     isotopic ratio such as 'C-12/C-13'.
-	yaxis : string
-	     Y axis, isotopic ratio such as 'C-12/C-13'.
-	solar_ab : string
-	     Path to the solar abundances used to normalize abundance ratios.
-	solar_iso : string
-	     Path to the solar isotopes used to normalize the isotopic ratios.
-        fig : string, float
-	     Figure name.
-        source : string
-            Specificies if yields come from
-	    all sources ('all') or from
-	    distinctive sources:
-	    only AGB stars ('agb'), only
-	    SN Ia ('sn1a'), or only massive stars
-	    ('massive')
-	marker : string
-	     Figure marker.
-	shape : string
-	     Line style.
-	color : string
-	     Line color.
-	label : string
-	     Figure label.
-	fsize : 2D float array
-	     Figure dimension/size.
-	fontsize : integer
-	     Font size of the numbers on the X and Y axis.
-	rspace : float
-	     Extra space on the right for the legend.
-	bspace : float
-	     Extra space at the bottom for the Y axis label.
-	labelsize : integer
-	     Font size of the X and Y axis labels.
-	legend_fontsize : integer
- 	       Font size of the legend.
-
-        Examples
-	----------
-
-	>>> o1.plot_iso_ratio(xaxis='[Fe/H]', yaxis='C-12/C-13')
-
-        '''
-
-        # Marker on the plots
-        markevery=500
-
-        # Declaration of the array to plot
-        x = []
-        y = []
-
-        #Access solar abundance
-        iniabu=ry.iniabu(global_path+solar_ab)
-        x_ini_iso=iniabu.iso_abundance(self.history.isotopes)
-        elements,x_ini=self._iso_abu_to_elem(x_ini_iso)
-
-        # Get the wanted source
-        if source == 'all':
-            yields_evol=self.history.ism_iso_yield
-            yields_evol_el=self.history.ism_elem_yield
-        elif source =='agb':
-            yields_evol=self.history.ism_iso_yield_agb
-            yields_evol_el=self.history.ism_elem_yield_agb
-        elif source == 'sn1a':
-            yields_evol=self.history.ism_iso_yield_1a
-            yields_evol_el=self.history.ism_elem_yield_1a
-        elif source == 'massive':
-            yields_evol=self.history.ism_iso_yield_massive
-            yields_evol_el=self.history.ism_elem_yield_massive
-        else:
-            print '!! Source not valid !!'
-            return
-
-        # Verify the X-axis
-        xaxis_ratio = False
-        if xaxis == 'age':
-            x = self.history.age
-        elif xaxis[0] == '[' and xaxis[-1] == ']':
-            xaxis_elem1 =xaxis.split('/')[0][1:]
-            xaxis_elem2 =xaxis.split('/')[1][:-1]
-            if not xaxis_elem1 in self.history.elements and \
-               not xaxis_elem2 in self.history.elements:
-                 print '!! Elements in xaxis are not valid !!'
-                 return
-
-            #X-axis ini values
-            x_elem1_ini=x_ini[elements.index(xaxis_elem1)]
-            x_elem2_ini=x_ini[elements.index(xaxis_elem2)]
-
-            #X-axis gce values
-            elem_idx1=self.history.elements.index(xaxis_elem1)
-            elem_idx2=self.history.elements.index(xaxis_elem2)
-
-            for k in range(0,len(yields_evol_el)):
-                if sum(yields_evol_el[k]) == 0:
-                    continue
-                #in case no contribution during timestep
-                x1=yields_evol_el[k][elem_idx1]/sum(yields_evol_el[k])
-                x2=yields_evol_el[k][elem_idx2]/sum(yields_evol_el[k])
-                spec=np.log10(x1/x2) - np.log10(x_elem1_ini/x_elem2_ini)
-                x.append(spec)
-        else:
-            xaxis_ratio = True
-            x_1 = ''
-            x_2 = ''
-            is_x_1 = True
-            for ix in range(0,len(xaxis)):
-                if xaxis[ix] == '/' and is_x_1:
-                    is_x_1 = False
-                else:
-                    if is_x_1:
-                        x_1 += xaxis[ix]
-                    else:
-                        x_2 += xaxis[ix]
-            if len(x_2) == 0:
-                print '!! xaxis not valid.  Need to be \'age\' or a ratio !!'
-                return
-
-        # Verify the Y-axis
-        y_1 = ''
-        y_2 = ''
-        is_y_1 = True
-        for iy in range(0,len(yaxis)):
-            if yaxis[iy] == '/' and is_y_1:
-                is_y_1 = False
-            else:
-                if is_y_1:
-                    y_1 += yaxis[iy]
-                else:
-                    y_2 += yaxis[iy]
-        if len(y_2) == 0:
-            print '!! yaxis not valid.  Need to be a ratio !!'
-            return
-
-        # Get the isotopes for X-axis
-        if xaxis_ratio:
-          if x_1 in self.history.isotopes and x_2 in self.history.isotopes:
-            idx_1=self.history.isotopes.index(x_1)
-            idx_2=self.history.isotopes.index(x_2)
-            x1_elem = str(x_1.split('-')[0])
-            x2_elem = str(x_2.split('-')[0])
-            x1_at_nb = float(x_1.split('-')[1])
-            x2_at_nb = float(x_2.split('-')[1])
-          else:
-            print '!! Isotopes in xaxis are not valid !!'
-            return
-
-        # Get the isotopes for Y-axis
-        if y_1 in self.history.isotopes and y_2 in self.history.isotopes:
-            idy_1=self.history.isotopes.index(y_1)
-            idy_2=self.history.isotopes.index(y_2)
-            y1_elem = str(y_1.split('-')[0])
-            y2_elem = str(y_2.split('-')[0])
-            y1_at_nb = float(y_1.split('-')[1])
-            y2_at_nb = float(y_2.split('-')[1])
-        else:
-            print '!! Isotopes in yaxis are not valid !!'
-            return
-
-        # Set the default label .. if not defined
-        if len(label)<1:
-            label=yaxis
-            if source=='agb':
-                label=yaxis+', AGB'
-            elif source=='massive':
-                label=yaxis+', Massive'
-            elif source=='sn1a':
-                label=yaxis+', SNIa'
-
-        # Get the solar values
-        if xaxis_ratio:
-            x1_sol, x2_sol, y1_sol, y2_sol = \
-              self.__read_solar_iso(solar_iso, x_1, x_2, y_1, y_2)
-        else:
-            x1_sol, x2_sol, y1_sol, y2_sol = \
-                self.__read_solar_iso(solar_iso, '', '', y_1, y_2)
-
-        # Calculate the isotope ratios (delta notation)
-        for k in range(0,len(yields_evol)):
-            if xaxis_ratio:
-                ratio_sample = (yields_evol[k][idx_1]/yields_evol[k][idx_2])*\
-                               (x2_at_nb / x1_at_nb)
-                ratio_std = x1_sol / x2_sol
-                x.append( ((ratio_sample/ratio_std) - 1) * 1000)
-            ratio_sample = (yields_evol[k][idy_1]/yields_evol[k][idy_2])*\
-                           (y2_at_nb / y1_at_nb)
-            ratio_std = y1_sol / y2_sol
-            y.append( ((ratio_sample/ratio_std) - 1) * 1000)
-  
-        # Make sure the length of array are the same when xaxis = '[X/Y]'
-        too_much = len(y)-len(x)
-        y = y[too_much:]
-
-        #Reserved for plotting
-        if not return_x_y:
-            shape,marker,color=self.__msc(source,shape,marker,color)
-
-        #Reserved for plotting
-        if not return_x_y:
-           plt.figure(fig, figsize=(fsize[0],fsize[1]))
-               #if xaxis == 'age':
-               #plt.xscale('log')
-           #plt.yscale('log')
-           plt.ylabel("$\delta$($^{"+str(int(y1_at_nb))+"}$"+y1_elem+"/$^{"+\
-                           str(int(y2_at_nb))+"}$"+y2_elem+")")
-           if xaxis == 'age':
-               plt.xlabel('Age [yr]')
-           elif xaxis_ratio:
-               plt.xlabel("$\delta$($^{"+str(int(x1_at_nb))+"}$"+x1_elem+"/$^{"+\
-                           str(int(x2_at_nb))+"}$"+x2_elem+")")
-           else:
-               plt.xlabel(xaxis)
-
-        #If x and y need to be returned ...
-        if return_x_y:
-            return x, y
-
-        else:
-            plt.plot(x,y,label=label,linestyle=shape,marker=marker,\
-                     color=color,markevery=markevery)
-            plt.legend()
-            ax=plt.gca()
-            self.__fig_standard(ax=ax,fontsize=fontsize,labelsize=labelsize,\
-                  rspace=rspace, bspace=bspace,legend_fontsize=legend_fontsize)
-	    #plt.xlim(self.history.dt,self.history.tend)	
-	    self.save_data(header=['Iso mass ratio',yaxis],data=[x,y])
-
-
-    ##############################################
-    #               Read Solar Iso               #
-    ##############################################
-    def __read_solar_iso(self, file_path, xx1, xx2, yy1, yy2):
-
-        # Initialisation of the variable to be returned
-        x1_rsi = -1
-        x2_rsi = -1
-        y1_rsi = -1
-        y2_rsi = -1
-
-        # Open the data file
-        with open(global_path+file_path, 'r') as data_file:
-
-            # For every line (for each isotope) ...
-            for line_1_str in data_file:
-
-                # Split the line
-                split = [str(x) for x in line_1_str.split()]
-
-                # Copy the number fraction of the isotopes (if found)
-                if split[0] == xx1:
-                    x1_rsi = float(split[1])
-                if split[0] == xx2:
-                    x2_rsi = float(split[1])
-                if split[0] == yy1:
-                    y1_rsi = float(split[1])
-                if split[0] == yy2:
-                    y2_rsi = float(split[1])
-
-        # Close the file
-        data_file.close()
-
-        # Return the number fractions
-        return x1_rsi, x2_rsi, y1_rsi, y2_rsi
-
-
-    ##############################################
-    #                  Plot MDF                  #
-    ##############################################
-    def plot_mdf(self, fig=19, return_x_y=False, axis_mdf = '[Fe/H]',\
-        dx = 0.05, solar_ab='yield_tables/iniabu/iniab2.0E-02GN93.ppn',\
-        sigma_gauss=0.0, nb_sigma = 3.0,\
-        marker='',shape='',\
-        color='',label='',fsize=[10,4.5],fontsize=14,rspace=0.6,\
-        bspace=0.15,labelsize=15,legend_fontsize=14):
-
-        '''
-	This function plots the stellar metallicity distribution function (MDF),
-	or the distribution function of any other abundance ratio.
-
-        Parameters
-        ----------
-
-	return_x_y : boolean
-	     If False, show the plot.  If True, return two arrays containing
-	     the X and Y axis data, respectively.
-	axis_mdf : string
-	     Abundance ratio for the distribution such as '[Fe/H]' or '[Mg/H]'.
-	dx : float
-	     Bin resolution of the distribution.
-	solar_ab : string
-	     Path to the solar abundances used to normalize abundance ratios.
-	sigma_gauss : float
-	     Each point in the MDF can be associated with a gaussian. This implies
-	     that in reality, the metallicity should have a certain dispersion 
-	     when stars form at each timestep (instead of using only a single
-	     average value).  The sigma_guass parameter sets the sigma value eac
-	     gaussian function.
-	nb_sigma : float
-	     When sigma_gauss is greater than zero, nb_sigma is the number of
-	     sigma by which the X axis range is expanded.
-        fig : string, float
-	     Figure name.
-	marker : string
-	     Figure marker.
-	shape : string
-	     Line style.
-	color : string
-	     Line color.
-	label : string
-	     Figure label.
-	fsize : 2D float array
-	     Figure dimension/size.
-	fontsize : integer
-	     Font size of the numbers on the X and Y axis.
-	rspace : float
-	     Extra space on the right for the legend.
-	bspace : float
-	     Extra space at the bottom for the Y axis label.
-	labelsize : integer
-	     Font size of the X and Y axis labels.
-	legend_fontsize : integer
- 	       Font size of the legend.
-
-        Examples
-	----------
-
-	>>> o1.plot_mdf(axis_mdf='[Fe/H]', sigma_gauss=0.2)
-
-        '''
-
-        # Define variable only used for plot visual
-        source='all'
-        markevery = 10000
-
-        # Get the [X/Y] values
-        x_min, x_max, x_all = self.__get_XY(axis_mdf, solar_ab)
-
-        # Construct the x axis for the MDF (lower boundary of the bin)
-        mdf_x = []
-        mdf_y = []
-        x_i = x_min
-        while x_i <= x_max:
-            mdf_x.append(x_i)
-            mdf_y.append(0.0)
-            x_i += dx
-        mdf_x.append(x_i)
-        mdf_y.append(0.0)
-
-        # For each step in the simulation ...
-        # len(x_all)-1 because we move by [low,up] using i_sim and i_sim+1
-        for i_sim in range(0, len(x_all)-1):
-
-          # Check if the values are ok
-          if np.isfinite(x_all[i_sim]) and np.isfinite(x_all[i_sim+1]):
-
-            # Calculate delta X / Mlocked
-            f_mlocked = self.history.m_locked[i_sim] / \
-              (x_all[i_sim+1] - x_all[i_sim])
-
-            # For each MDF x bin ...
-            # len(mdf_x)-1 because mdf_x[-1] = last upper lim
-            for i in range(0, len(mdf_x)-1):
-
-              # If there is an overlap, add the corresponding mass in the bin
-              # Bin is overlaping the lower-boundary of the simulation
-              if mdf_x[i] <= x_all[i_sim] and \
-                 x_all[i_sim] < mdf_x[i+1] and \
-                 mdf_x[i+1] < x_all[i_sim+1]:
-                  mdf_y[i] += (mdf_x[i+1] - x_all[i_sim]) * f_mlocked
-
-              # Bin is totaly within the boundarie of the simulation 
-              elif x_all[i_sim] <= mdf_x[i] and \
-                mdf_x[i+1] < x_all[i_sim+1]:
-                  mdf_y[i] += (mdf_x[i+1] - mdf_x[i]) * f_mlocked 
-
-              # Bin is overlaping the uper-boundary of the simulation
-              elif x_all[i_sim] <= mdf_x[i] and \
-                mdf_x[i] < x_all[i_sim+1] and \
-                x_all[i_sim+1] < mdf_x[i+1]:
-                  mdf_y[i] += (x_all[i_sim+1] - mdf_x[i]) * f_mlocked
-
-              # Bin is totaly overlaping the simulation
-              elif mdf_x[i] <= x_all[i_sim] and \
-                x_all[i_sim+1] < mdf_x[i+1]:
-                  mdf_y[i] += (x_all[i_sim+1] - x_all[i_sim]) * f_mlocked
-
-        # Correct the x axis to use the middle value of each bin
-        x_cor = dx/2.0
-        for i in range(0, len(mdf_x)):
-            mdf_x[i] += x_cor
-
-        # Normalisation to 1.0 (max value)
-        cte = 1.0 / max(mdf_y)
-        for i in range(0, len(mdf_y)):
-            mdf_y[i] = mdf_y[i] * cte
-
-        # If the gaussian approximation is wanted ...
-        if sigma_gauss > 0.0:
-            mdf_x, mdf_y = \
-                self.__gauss_approx(mdf_x, mdf_y, sigma_gauss, dx, nb_sigma)
-
-        # Reserved for plotting
-        if not return_x_y:
-            shape,marker,color=self.__msc(source,shape,marker,color)
-            plt.figure(fig, figsize=(fsize[0],fsize[1]))
-            plt.xlabel(axis_mdf)
-            plt.ylabel('$N_\star$')
-            plt.plot(mdf_x,mdf_y,label=label,linestyle=shape,marker=marker,\
-                     color=color,markevery=markevery)
-            plt.legend()
-            ax=plt.gca()
-            self.__fig_standard(ax=ax,fontsize=fontsize,labelsize=labelsize,\
-                  rspace=rspace, bspace=bspace,legend_fontsize=legend_fontsize)
-
-        # If the MDF should be returned ...
-        else :
-            return mdf_x, mdf_y
-
-
-    ##############################################
-    #                 Get [X/Y]                  #
-    ##############################################
-    def __get_XY(self, axis_mdf = '[Fe/H]', \
-             solar_ab='yield_tables/iniabu/iniab2.0E-02GN93.ppn'):
-
-        # Access solar abundance
-        iniabu = ry.iniabu(global_path+solar_ab)
-        x_ini_iso = iniabu.iso_abundance(self.history.isotopes)
-
-        # Access 
-        elements, x_ini = self._iso_abu_to_elem(x_ini_iso)
-        yields_evol = self.history.ism_elem_yield
-
-        # Isolate the X and Y in [X/Y]
-        xaxis_elem1 = axis_mdf.split('/')[0][1:]
-        xaxis_elem2 = axis_mdf.split('/')[1][:-1]
-
-        # Get the solar values
-        x_elem1_ini = x_ini[elements.index(xaxis_elem1)]
-        x_elem2_ini = x_ini[elements.index(xaxis_elem2)]
-
-        # Get the simulation values
-        elem_idx1 = self.history.elements.index(xaxis_elem1)
-        elem_idx2 = self.history.elements.index(xaxis_elem2)
-
-        x = []
-        x_no_inf = []
-        # Calculate [X/Y] at each timestep
-        for k in range(0,len(yields_evol)):
-            x1 = yields_evol[k][elem_idx1] / sum(yields_evol[k])
-            x2 = yields_evol[k][elem_idx2] / sum(yields_evol[k])
-            spec = np.log10(x1/x2) - np.log10(x_elem1_ini/x_elem2_ini)
-            if np.isfinite(spec):
-                x_no_inf.append(spec)
-            x.append(spec)
-
-        # Return [X/Y]_min, [X/Y]_max, and the [X/Y] array
-        x_no_inf = np.sort(x_no_inf)
-        return x_no_inf[0], x_no_inf[-1], x
-
 
     ##############################################
     #               Gauss Approx.                #
